Init commit of First Frontend application.
This frontend application allows the end user to set up an account, answer basic questions, and then see a dashboard where they can log their data. the initial state is that it doesn't have a DB set up, but should be relatively easy to drop one in later, and then also the LLM stuff is not in either.